Separate Done Button as a component in DoneButton.tsx
